Add openSUSE support to devel role
Provide an empty kheader task file for now to at least avoid an
"unsupported os" error when running the devel role on a openSUSE 15
instance.

Add the proper kernel devel packages to build kernel modules and remove
the work-around tasks to fiddle with the kernel headers.

Skip opensuse package tests in devel role