Refactor command line assembly in openafs_build module
Move the code that assemble the configure and make command lines into
functions, and update the module documentation.

This change should not make any functional changes, but reorganizes the
code in preparation for more openafs_build options.  Importantly, this
change allows the module to detect when the configure_options is
omitted.