Add build_* and with_* options to openafs_build module
Add a set of build options to the openafs_build module to determine the
configure and make arguments when those are not explicitly specified.

Adopt the following convention for build options:

    build_*  specifies what to build
    with_*   specifies options within the built programs

Rename some existing options to follow this convention, but use
module parameter aliases for backward compatibility.

Remove explicit configure_options in the roles.

Ignore and undocument the 'state' parameter. It was rather unless in
this module, since this module is really process driver and not a state
change to the node.