modules: Detect installation directories
Instead of assuming the installation directories based on the type of
package manager, check the installed man pages and look for the
configuration directory paths.  Hopefully in the future, the programs
will have an option to reveal configuration directories.

Also, read the cacheinfo file installed by the packaging to set the
cacheinfo local facts.

This requires the doc packages to also be installed, but the roles
already install them by default. If you need to avoid installing the doc
package, you'll need to provide the config dirs manually.