Detect AFSD_ARGS installed by packages
Instead of clobbering the AFSD_ARGS installed by packages, detect and
save them in the local facts file, and use them as the default for the
afs_afsd_opts client role variable. That variable can be used as an
inventory variable to override the packaged defined AFSD_ARGS to set
specific afsd args.

Currently, only systemd client startup is supported for this AFSD_ARGS
detection.  Try to file the AFSD_ARGS in the systemd EnvironmentFile
specified by the openafs-client service unit file. If not found, fall
back to the default values we use when installing from source.

The local facts are now set based on the cacheinfo installed from the
packaging. Use those values for the cache config parameters instead of
hardcoded values based on the distro type.