client: Unlink client CellServDB from server CellServDB
Unfortunately, the bosserver creates a client side CellServDB file as a
symlink to the server side CellServDB file (unless the client side file
already exists).  So, when the client role assembles a client side
CellServDB file on a node that is a combined server and client, the
server side CellServDB is overwritten with the CellServDB.dist records.

Be sure to remove the symlink before updating the client CellServDB
file.