modules: Do not clutter syslog with local facts
Currently, a poorly formatted and very large syslog message consisting
of the full set of local facts is written to the syslog each time a
local fact is stored.

Avoid cluttering the syslog by only dumping the local facts as a debug
log entry, and remove the newline and extra whitespace, which do not
reader well in the syslog.