Adding environment with values from configure_env variable
When building from source it is possible to define variables
to fine-tune directories. Directory variables which start with afs
cannot be configured via passing flags to configure command.

For example:
  afslocaldir=SOMEPATH
  afslogsdir=ANOTHERPATH

Note: default(omit) will produce a warning if configure_env is not
defined:
[WARNING]: could not parse environment value, skipping: ['{{ configure_env |
default(omit) }}']

With default({}) no warning will be created.