Add support for extra environment variables during configure
Add a module parameter to the openafs_build module to specify extra
configure variables when configure is run.

Add new roles variables to set the configure environment when installing
from source (and the biniaries are built by the roles).

This allows us to configure the tree to use customized paths, for
example to configure a tree with a customized logs path by setting the
'afslogsdir' environment variable.  See src/cf/dirpath.cf in the OpenAFS
source tree for paths that can be configured this way.