Set the fallback afs_user to 'ansible'
The username running the play is used to set the afs_user and afs_admin
role variable defaults. If these variables are not set, and the username
cannot be determined, fallback to 'ansible' and 'ansible.admin'.

Clean up the realm.yaml example playbook. Do not hard code 'test' as the
user principal, and instead use the value of afs_user, which matches the
principal name in the cell.yaml example playbook.  Also remove the extra
openafs_common role import, since that is now done by the
openafs_krbserver role.