Probe for client dynroot mode
The process for mounting the root.cell volume depends on the client's
dynroot mode, so the openafs_volume module currently has a check to see
if dynroot is enabled or not before mounting the root.cell volume.

Currently, the check for dynroot mode is done by searching for the
-dynroot and -fakestat (and variants arguments). However, that method is
error prone and complicated.  Originally, the afsd args used were always
given by the openafs_client afs_afsd_opts role variable, but when
installing from packages, we use the packaged values when afs_afsd_opts
is not defined.

Instead of checking what we think are the current afsd arguments, probe
for the dynroot mode by stating a dynamically mounted path of the root
vnode of the root.cell volume (which must exist before the client is
started).  If that path stat fails with ENODEV, we know the client is in
legacy mode (that is dynroot is not enabled).