Remove afs_afsd_opts from the client role defaults
Originally, afs_afsd_opts varialbe was always used to set the afsd
command line args, but later we changed the client role to use the args
provide by the installation packages by default, which makes more sense.
When that change was made, code was added to detect the afsd options
after the packages were installed and use that to set the afs_afsd_opts,
but that is complicated and error prone.

Also, the afs_afsd_opts was also orignally used to determine if dynroot
was enabled, which was needed to mount the root cell volume.  That check
was changed to probe from the dynroot mode, so the afsd args are no
longer needed to mount the root cell volume.

Clean this up by removing afs_afsd_opts from the role defaults, so that
variable is undefined by default.  If is defined, then override the
packaged value in the platform specific configuration file.

When installing from source, define the afsd args default in the
templates used to configure the startup service.  These value will be
overriden if an afs_afsd_opts inventory variable is defined.