Use local facts only for the service names
The client and server service names are stored in local facts, but we
currently allow that to be overridden by role variables.  But the
service names are packaging dependent, so always use the local fact
directly for the service name.

If the local fact is not present, then the service has not been setup
yet, so in some cases were we need to shutdown a service to update, we
check the local fact is defined to determine if the service has been
setup.