Refactor the openafs_get_install_paths module
Rework the package manager autodetection by calling the Ansible
module_utils PkgMgrFactCollector class to find the system package
manager.  (This class used to find the ansible_pkg_mgr fact in the
standard setup module.)

Rename the old PkgMgr base class to InstallationFactCollector, to be in
line with the naming convention for collectors.  Provide a factory
factory method in the base class (by a __new__ override) to instantiate
a suitable sub-class for package manager specific processing.

Use the standard module.run_command() method to run the rpm and
dpkg-query commands.  Refactor the code between the base case and the
subclasses so the platform specific command processing is less
convoluted.

Move the various platform agnostic helper functions to the base clase.