Refactor the openafs_build_sdist module
Use the standard run_command() module utility function to run commands
instead of a custom subprocess wrapper.

Remove globals and group the functions into a class.  Use a factory
method for handling platform specific code.  Move the solaris specific
tranlation of 'tar' to 'gtar' to the solaris subclass.

Move the compression routine to a class. Run the compressions only for
the ones found, and create the checksum only if the md5sum program is
found.  Add a table of supported compression formats at the top of the
file, to make it easier to add new ones in the future. The Compression
class can be subclasses to support new compression types.