krbserver: Remove afs_krb_admin_password variable
Use use a keytab for the admin user, so it does not make sense to also
set the interactive password.  Remove the password variable from the
krbserver role.

Not only is this more secure, since it is one less place to store
a password, but it also let's us change the openafs_principal module
to only create keytabs when the password is not specified.

This is an incompatible change if you were using an interactive password
for the admin user. If that was the case you should create a different
principal for interactive use.