modules: Improve not implemented error messages
Improve the error messages if an unimplemented method is called in the
openafs_principal module.  If method is on the base class, it is because
a platform was not found. If the method is on a subclass, then throw an
exception, because the base class is incomplete.

Also add missing kadmin_args() abstract method.

Also fix missing password argument.