roles: Refactor cancel restart handler
Use the recommended method to cancel the service restart notification,
by setting a fact after the service started task to cancel the restart
in the handlers.

Rename the notifications to "(server|client) configuration changed", and
add the listen property in the hanlder task.

Remove the local fact contition from the restart handler, since we
always use the service manager on all supported platforms now.

The restart can be also skipped by setting the
afs_cancel_restart_*_service variable in your inventory or when the
roles are imported.