roles: Add afs_devel_install_tools role variable
Add a role variable to the openafs_devel role so we can install the
kernel headers and devel tools separately.

If we are installing the kernel headers (afs_devel_kernel_headers is
true), but not installing the full set of development tools, be sure to
install the tools needed to build the kernel module.