roles: Install kernel module build requirements
When building the the client, either from source or building the DMKS
module, import the openafs_devel role to import the kernel headers
matching our running kernel version and build tools required to build a
kernel module.  To do this, replace the tasks to install various
packages with an openafs_devel role import.

The openafs_devel role may install a new kernel version and a reboot to
get a kernel header package. If you do not want to updrade and reboot,
set the afs_devel_allow_kernel_update role variable to false for the
given host.

When building only the userspace binaries on the server, avoid
installing the kernel headers (and potentially updating the kernel and
rebooting).