roles: Platform dependent afs_install_method default
Instead of always hardcoding "managed" as the default value for
afs_install_method, pick a sensible value depending on the platform.

Currently, the linux platforms default to "managed", and the FreeBSD and
Solaris platforms default to "source".

Use the platform default in the default molecule roles, so when running
a test we can just pick an image name and let the platform type pick the
default install method for that platform.

Treat an empty afs_install_method string as undefined. This is a bit
trickery, since we cant jut set the default value in the
defaults/main.yaml, but it let's us set a env var defined value in
molecule, which sets the variable to empty when the env var is unset.