roles: Treat empty afs_git_* variables as undefined
Sadly, the role variables are set to empty values when an environment
variable is used in molecule and the enviroment variable is not set and
does not have a default value. This makes it impossible to test the
role's default value.

To work around this, use the default(x, true) filter where the afs_git_*
role variables are used, so empty vars are handled like undefined values
and we pick a default.