modules: Detect git clean exclude option
The git clean --exclude option is not available in very old versions of
git, notably, it is not available in the git that ships in
RHEL6/CentOS6.

Add a git clean usage check to see if the --exclude option is available,
and if not, avoid running 'git clean' (since we do not want to clobber our
logs), and fallback to running 'make clean'.