roles: Install autoconf 2.69 on RHEL/CentOS 6
The autoconf that ships with RHEL/CentOS 6 is too old to generate the
OpenAFS configure script.  Install a usable version from a source
download.

Since this needs to be done in both the devel and devel_kernel roles,
for RHE6 and CentOS6 in each, put the autoconf tasks in the common role
and import the tasks where needed.