Dynamic afs_cachedir default
Currently, the afs_cachedir always defaults to "/usr/vice/cache" when
building from source.  However, we really should be using the modern
"/var/cache/openafs" path when installing with modern paths.

Instead, dynamically determine the afs_cachedir default value when
building from source.  When the viceetcdir results from the build is
/usr/vice/etc, use "/usr/vice/cache" as the afs_cachedir default,
otherwise, use the modern path "/var/cache/openafs" for the afs_cachedir
default.

If the afs_cachedir is defined as an inventory variable, the inventory
variable will override the default value, regardless of the results of
the build.

Update the server role build/all.yam tasks as well, for the situation
where the node is both a client and a server.

Also, to be consistent with the package install methods, save the
mountpoint, cachedir, cachesize as a local facts when installing the
client from source. The values are saved in the
"/etc/ansible/facts.d/openafs.fact" file on the client node.