modules: Deprecate redundant build parameters
In practice, we always need to specify the configure option, so
deprecate the "what to build" options in the openafs_build module.  Log
a warning for now, but in the future these will be removed.

This also avoids needing to change the build module as new options are
added to configure.