roles: Add remote_src to unarchive tasks
Add the remote_src to all of the unarchive tasks to automatically
download the release tarball from a URL when it contains "://" in the
path.  Set remote_src to be true by default, so we the release tarball
is assumed to be on the remote node, to make it easier to build
candidate RPMs on the test machine then test them.

Add the afs_install_archive_remote_src variable to override this, in
case you want to upload the tarball from the controller.