tests: Rewrite pytest frontend to molecule
Rewrite the pytests to run module to be less janky.

Use a regular pytest fixture to run molecule and avoid the external
custom pytest.skip files.  Add regular pytest functions to run molecule
for each role and for the playbooks.  Use tox to create the venv and run
pytest.

Run black on the files under tests to standarize the formatting.